requirements.txt
streamlit>=1.34
pandas>=2.2
plotly>=5.22
numpy>=1.26
openpyxl>=3.1
Pillow>=10.3
requests>=2.32
